Move classes to duke package
